+++ trionyx/forms/layout.py
@@ -118,3 +118,17 @@ class TimePicker(DateTimePicker):
     format = '%H:%M'
     show_today_button = False
     glyphicon = 'glyphicon-time'
+
+
+class Formset(LayoutObject):
+    template = 'trionyx/forms/formset.html'
+
+    def __init__(self, formset_name_in_context, template=None):
+        self.formset_name_in_context = formset_name_in_context
+        self.fields = []
+        if template:
+            self.template = template
+
+    def render(self, form, form_style, context, template_pack=TEMPLATE_PACK):
+        formset = context[self.formset_name_in_context]
+        return render_to_string(self.template, {'formset': formset})